backend/backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
shortuuid==1.0.13
sqlparse==0.5.3
uritemplate==4.1.1
uvicorn==0.34.0
whitenoise==6.9.0
